+++ packages/backend/api/src/types.ts
@@ -2,4 +2,4 @@ export type * from './game/get.zones-count';
 export type * from './game/barrier/get.barriers-route';
 export type * from './game/get.enclosures';
 export type * from './game/get.decorations';
-export type * from './game/get.creatures-route';
+export type * from './game/get.creatures';
