+++ packages/frontend/web/src/components/create-park.tsx
@@ -5,7 +5,7 @@ import { useAuth } from '@/contexts/auth-context';
 import { useGameInfoContext } from '@/contexts/game-info-context';
 
 import ButtonBlue from './button-blue';
-import InputBlue from './input-blue';
+import Input from './input';
 
 const API_URL = import.meta.env.VITE_API_URL;
 
@@ -67,7 +67,9 @@ export default function CreatePark() {
           await createPark();
         }}
       >
-        <InputBlue
+        <Input
+          bgColor='bg-primary-blue'
+          borderColor='border-secondary-blue'
           type='text'
           placeholder='Park name'
           value={parkName}
