+++ packages/frontend/web/src/components/creature-line.tsx
@@ -1,28 +1,9 @@
 import useCreatures from '@/hooks/use-creatures';
+import { remainingTimeToFeed } from '@/hooks/use-format-time-until';
 
 import Female from '../assets/images/icons-buttons/female.png';
 import Male from '../assets/images/icons-buttons/male.png';
 import ButtonBuy from './button-buy';
-
-function remainingTimeToFeed(ComingDate: Date) {
-  const now = new Date();
-  const diff = ComingDate.getTime() - now.getTime();
-
-  if (diff <= 0) return 'Feed me !';
-
-  const totalMins = Math.floor(diff / 1000 / 60);
-  const days = Math.floor(totalMins / 1440);
-  const hours = Math.floor((totalMins % 1440) / 60);
-  const minutes = totalMins % 60;
-
-  const remainingTime = [];
-  if (days > 0) remainingTime.push(`${days} day${days > 1 ? 's' : ''}`);
-  if (hours > 0) remainingTime.push(`${hours} hour${hours > 1 ? 's' : ''}`);
-  if (minutes > 0 && days === 0)
-    remainingTime.push(`${minutes} min${minutes > 1 ? 's' : ''}`);
-
-  return ` ${remainingTime.join(' ')}`;
-}
 
 function getPotionImage(zoneId: number) {
   switch (zoneId) {
