+++ packages/frontend/web/src/hooks/use-format-time-until.ts
@@ -1,0 +1,19 @@
+export function remainingTimeToFeed(ComingDate: Date) {
+  const now = new Date();
+  const diff = ComingDate.getTime() - now.getTime();
+
+  if (diff <= 0) return 'Feed me !';
+
+  const totalMins = Math.floor(diff / 1000 / 60);
+  const days = Math.floor(totalMins / 1440);
+  const hours = Math.floor((totalMins % 1440) / 60);
+  const minutes = totalMins % 60;
+
+  const remainingTime = [];
+  if (days > 0) remainingTime.push(`${days} day${days > 1 ? 's' : ''}`);
+  if (hours > 0) remainingTime.push(`${hours} hour${hours > 1 ? 's' : ''}`);
+  if (minutes > 0 && days === 0)
+    remainingTime.push(`${minutes} min${minutes > 1 ? 's' : ''}`);
+
+  return ` ${remainingTime.join(' ')}`;
+}
